backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1

# Data Processing
pandas==2.2.3
numpy==2.1.3
scipy==1.14.1

# Graph Database
neo4j==5.26.0

# Machine Learning - PyTorch
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1

# PyTorch Geometric
torch-geometric==2.6.1

# HTTP Requests
httpx==0.27.2
requests==2.32.3
aiohttp==3.11.7

# LLM Integration
google-generativeai==0.8.3

# Utilities
python-multipart==0.0.17

# Uvicorn dependencies (manual install to avoid watchfiles issue on Windows)
websockets>=10.0
httptools>=0.5.0
